pyproject.toml
[tool.poetry]
name = "langchain-practice"
version = "0.1.0"
description = "LangChainの使い方学習用プロジェクト"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.1.9"
langchain-community = "^0.0.21"
python-dotenv = "^1.0.0"
fastapi = "^0.105.0"
uvicorn = "^0.24.0"
pydantic = "^2.5.2"
chromadb = "^0.4.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 